login: only include relevant data layer handlers
I moved the 'registerHandlers' up a layer to 'state'. I don't think it's an ideal solution but I needed a way to not include all of wpcom data layer handlers in the login dependency tree. We can likely move it down again as soon as all data layer handlers are loaded dynamically.